Return everything as dicts instead of SQLAResult, Remove SQLAResult
Eve exspects to be able to change fields of returned objects. Returning
them as SQLAResult objects creates problems whenever the objects are
changed and changed are forwarded to the database, which is not the
exspected behaviour. This behaviour broke hooks, which wanted to returned
the original object after a PATCH operation.

See issue #11 at https://github.com/RedTurtle/eve-sqlalchemy/issues/11

This also fixes that sometimes objects where embedded when they should
not have. We have no more need for seperate jsonify functions, which
were not respecting embedding parameters.

Also the calculation of etags sometimes had problems with SQLAResult
objects, as the custom serializer was not used there. This is also fixed
now.