add support for PostgreSQL JSON type
With this change, it's now possible to push JSON in a JSON field. For
example, in this example, the struct file has the type JSON.

```
POST http://127.0.0.1:5000/notifications
Content-Type: application/json

[{"environment_id":"ae7d4ed8-41b4-0206-6aad-0074c0f5b6f1","struct":{"type":"gerrit","server":"somewhere","account":"bob","port":29418}}]
```

Without this change, the validator will fail because it don't know about
JSON. It will expect a string instead, which is the default type.