PUT must drop/recreate item in the same transaction
So far, PUT was calling replace() to drop, commit, recreate the item.
By doing this, we can get the situation where the initial drop break
the foreign keys of the other tables.
Instead, if we do the drop/recreate in the same transaction, the
constraint engine will ignore the unstable situation.

This commit extend the code coverage:
- enable SQLite FK constrain check. We use 'pragma foreign_keys=on'
  to do that.
- create a new notes table with a FK on the people._id key.
- closes #64.

See: https://www.sqlite.org/pragma.html