Avoid nasty import conflict
From the Python docs:
"When a submodule is loaded using any mechanism [...] a binding is
placed in the parent module’s namespace to the submodule object. For
example, if package spam has a submodule foo, after importing
spam.foo, spam will have an attribute foo which is bound to the
submodule."
(see https://docs.python.org/3/reference/import.html#submodules)

Since creating `eve_sqlalchemy.config` we ran in exactly this problem
as an import of the new module results in overriding the already
existing global `config` object from eve.utils.