Eve 0.7 support
The necessary premise is that I have (very) limited knowledge of
eve-sqlalchemy codebase. These changes might not be sufficient for
a Eve 0.7 compatible release of the eve-sqlalchemy extension.

I wanted to see what was preventing eve-sqlalchemy from running on Eve
0.7. As it turns out, there are no major obstacles. With this commit
all tests are passing with Eve<0.8, locally at least.

Next up would probably be a careful review of Eve 0.7 changelog, to see
what new features (or fixes) need to be specifically addressed here.